+++ libs/utils/index.ts
@@ -1,5 +1,11 @@
 export { EndowedAccounts } from "@/libs/utils/EndowedAccounts";
 export { errMsgFilter } from "@/libs/utils/errorHandling";
-export { default as fetchValidAccount } from "@/libs/utils/fetchValidAccount";
 export { default as supplyAccount } from "@/libs/utils/supplyAccount";
+export { default as fetchValidAccount } from "@/libs/utils/fetchValidAccount";
+export { default as addCENNZnetToMetaMask } from "@/libs/utils/addCENNZnetToMetaMask";
+export { default as addCENNZTokenToMetaMask } from "@/libs/utils/addCENNZTokenToMetaMask";
+export { default as cvmToCENNZAddress } from "@/libs/utils/cvmToCENNZAddress";
+export { default as ensureEthereumChain } from "@/libs/utils/ensureEthereumChain";
+export { default as isCENNZAddress } from "@/libs/utils/isCENNZAddress";
+export { default as isEthereumAddress } from "@/libs/utils/isEthereumAddress";
 export { default as trackPageView } from "@/libs/utils/trackPageView";