+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [1.0.1] - 2022-04-14
 
- - Alert user if their account is not valid [#29](https://github.com/cennznet/app-faucet/pull/29)
+- Alert user if their account is not valid [#29](https://github.com/cennznet/app-faucet/pull/29)
 
 ## [1.0.0] - 2022-04-13